Move opencensus receiver
