+++ receiver/opencensusreceiver/internal/ocmetrics/opencensus.go
@@ -28,7 +28,8 @@ import (
 	"go.opentelemetry.io/collector/config"
 	"go.opentelemetry.io/collector/consumer"
 	"go.opentelemetry.io/collector/obsreport"
-	"go.opentelemetry.io/collector/translator/internaldata"
+
+	internaldata "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/translator/opencensus"
 )
 
 // Receiver is the type used to handle metrics from OpenCensus exporters.
