+++ receiver/opencensusreceiver/internal/ocmetrics/opencensus_test.go
@@ -12,40 +12,43 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package octrace
+package ocmetrics
 
 import (
 	"bytes"
 	"context"
 	"encoding/json"
+	"fmt"
 	"io"
 	"net"
 	"strings"
+	"sync"
 	"testing"
 	"time"
 
 	commonpb "github.com/census-instrumentation/opencensus-proto/gen-go/agent/common/v1"
-	agenttracepb "github.com/census-instrumentation/opencensus-proto/gen-go/agent/trace/v1"
-	tracepb "github.com/census-instrumentation/opencensus-proto/gen-go/trace/v1"
+	agentmetricspb "github.com/census-instrumentation/opencensus-proto/gen-go/agent/metrics/v1"
+	metricspb "github.com/census-instrumentation/opencensus-proto/gen-go/metrics/v1"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc"
 	"google.golang.org/grpc"
 	"google.golang.org/protobuf/proto"
+	"google.golang.org/protobuf/types/known/timestamppb"
 
 	"go.opentelemetry.io/collector/component/componenttest"
 	"go.opentelemetry.io/collector/config"
 	"go.opentelemetry.io/collector/consumer"
 	"go.opentelemetry.io/collector/consumer/consumertest"
-	"go.opentelemetry.io/collector/exporter/opencensusexporter"
-	"go.opentelemetry.io/collector/internal/testdata"
-	"go.opentelemetry.io/collector/translator/internaldata"
+
+	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/opencensusexporter"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal/testdata"
+	internaldata "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/translator/opencensus"
 )
 
 func TestReceiver_endToEnd(t *testing.T) {
-	spanSink := new(consumertest.TracesSink)
+	metricSink := new(consumertest.MetricsSink)
 
-	addr, doneFn := ocReceiverOnGRPCServer(t, spanSink)
+	addr, doneFn := ocReceiverOnGRPCServer(t, metricSink)
 	defer doneFn()
 
 	expFactory := opencensusexporter.NewFactory()
@@ -53,7 +56,7 @@ func TestReceiver_endToEnd(t *testing.T) {
 	expCfg.GRPCClientSettings.TLSSetting.Insecure = true
 	expCfg.Endpoint = addr.String()
 	expCfg.WaitForReady = true
-	oce, err := expFactory.CreateTracesExporter(context.Background(), componenttest.NewNopExporterCreateSettings(), expCfg)
+	oce, err := expFactory.CreateMetricsExporter(context.Background(), componenttest.NewNopExporterCreateSettings(), expCfg)
 	require.NoError(t, err)
 	err = oce.Start(context.Background(), componenttest.NewNopHost())
 	require.NoError(t, err)
@@ -62,15 +65,15 @@ func TestReceiver_endToEnd(t *testing.T) {
 		require.NoError(t, oce.Shutdown(context.Background()))
 	}()
 
-	td := testdata.GenerateTracesOneSpan()
-	assert.NoError(t, oce.ConsumeTraces(context.Background(), td))
+	md := testdata.GenerateMetricsOneMetric()
+	assert.NoError(t, oce.ConsumeMetrics(context.Background(), md))
 
 	assert.Eventually(t, func() bool {
-		return len(spanSink.AllTraces()) != 0
+		return len(metricSink.AllMetrics()) != 0
 	}, 10*time.Second, 5*time.Millisecond)
-	gotTraces := spanSink.AllTraces()
-	require.Len(t, gotTraces, 1)
-	assert.Equal(t, td, gotTraces[0])
+	gotMetrics := metricSink.AllMetrics()
+	require.Len(t, gotMetrics, 1)
+	assert.Equal(t, md, gotMetrics[0])
 }
 
 // Issue #43. Export should support node multiplexing.
@@ -78,16 +81,16 @@ func TestReceiver_endToEnd(t *testing.T) {
 // a passthrough mode where it initiates Export normally by firstly
 // receiving the initiator node. However ti should still be able to
 // accept nodes from downstream sources, but if a node isn't specified in
-// an exportTrace request, assume it is from the last received and non-nil node.
+// an exportMetrics request, assume it is from the last received and non-nil node.
 func TestExportMultiplexing(t *testing.T) {
-	spanSink := new(consumertest.TracesSink)
+	metricSink := new(consumertest.MetricsSink)
 
-	port, doneFn := ocReceiverOnGRPCServer(t, spanSink)
+	addr, doneFn := ocReceiverOnGRPCServer(t, metricSink)
 	defer doneFn()
 
-	traceClient, traceClientDoneFn, err := makeTraceServiceClient(port)
-	require.NoError(t, err, "Failed to create the gRPC TraceService_ExportClient: %v", err)
-	defer traceClientDoneFn()
+	metricsClient, metricsClientDoneFn, err := makeMetricsServiceClient(addr)
+	require.NoError(t, err, "Failed to create the gRPC MetricsService_ExportClient: %v", err)
+	defer metricsClientDoneFn()
 
 	// Step 1) The initiation.
 	initiatingNode := &commonpb.Node{
@@ -98,58 +101,58 @@ func TestExportMultiplexing(t *testing.T) {
 		LibraryInfo: &commonpb.LibraryInfo{Language: commonpb.LibraryInfo_JAVA},
 	}
 
-	err = traceClient.Send(&agenttracepb.ExportTraceServiceRequest{Node: initiatingNode})
+	err = metricsClient.Send(&agentmetricspb.ExportMetricsServiceRequest{Node: initiatingNode})
 	require.NoError(t, err, "Failed to send the initiating message: %v", err)
 
-	// Step 1a) Send some spans without a node, they should be registered as coming from the initiating node.
-	sLi := []*tracepb.Span{{TraceId: []byte("1234567890abcdef"), Status: &tracepb.Status{}}}
-	err = traceClient.Send(&agenttracepb.ExportTraceServiceRequest{Node: nil, Spans: sLi})
+	// Step 1a) Send some metrics without a node, they should be registered as coming from the initiating node.
+	mLi := []*metricspb.Metric{makeMetric(1)}
+	err = metricsClient.Send(&agentmetricspb.ExportMetricsServiceRequest{Node: nil, Metrics: mLi})
 	require.NoError(t, err, "Failed to send the proxied message from app1: %v", err)
 
-	// Step 2) Send a "proxied" trace message from app1 with "node1"
+	// Step 2) Send a "proxied" metrics message from app1 with "node1"
 	node1 := &commonpb.Node{
 		Identifier:  &commonpb.ProcessIdentifier{Pid: 9489, HostName: "nodejs-host"},
 		LibraryInfo: &commonpb.LibraryInfo{Language: commonpb.LibraryInfo_NODE_JS},
 	}
-	sL1 := []*tracepb.Span{{TraceId: []byte("abcdefghijklmnop"), Name: &tracepb.TruncatableString{Value: "test"}, Status: &tracepb.Status{}}}
-	err = traceClient.Send(&agenttracepb.ExportTraceServiceRequest{Node: node1, Spans: sL1})
+	mL1 := []*metricspb.Metric{makeMetric(2)}
+	err = metricsClient.Send(&agentmetricspb.ExportMetricsServiceRequest{Node: node1, Metrics: mL1})
 	require.NoError(t, err, "Failed to send the proxied message from app1: %v", err)
 
-	// Step 3) Send a trace message without a node but with spans: this
+	// Step 3) Send a metrics message without a node but with metrics: this
 	// should be registered as belonging to the last used node i.e. "node1".
-	sLn1 := []*tracepb.Span{{TraceId: []byte("ABCDEFGHIJKLMNOP"), Status: &tracepb.Status{}}, {TraceId: []byte("1234567890abcdef"), Status: &tracepb.Status{}}}
-	err = traceClient.Send(&agenttracepb.ExportTraceServiceRequest{Node: nil, Spans: sLn1})
+	mLn1 := []*metricspb.Metric{makeMetric(3)}
+	err = metricsClient.Send(&agentmetricspb.ExportMetricsServiceRequest{Node: nil, Metrics: mLn1})
 	require.NoError(t, err, "Failed to send the proxied message without a node: %v", err)
 
-	// Step 4) Send a trace message from a differently proxied node "node2" from app2
+	// Step 4) Send a metrics message from a differently proxied node "node2" from app2
 	node2 := &commonpb.Node{
 		Identifier:  &commonpb.ProcessIdentifier{Pid: 7752, HostName: "golang-host"},
 		LibraryInfo: &commonpb.LibraryInfo{Language: commonpb.LibraryInfo_GO_LANG},
 	}
-	sL2 := []*tracepb.Span{{TraceId: []byte("_B_D_F_H_J_L_N_O"), Status: &tracepb.Status{}}}
-	err = traceClient.Send(&agenttracepb.ExportTraceServiceRequest{Node: node2, Spans: sL2})
+	mL2 := []*metricspb.Metric{makeMetric(4)}
+	err = metricsClient.Send(&agentmetricspb.ExportMetricsServiceRequest{Node: node2, Metrics: mL2})
 	require.NoError(t, err, "Failed to send the proxied message from app2: %v", err)
 
-	// Step 5a) Send a trace message without a node but with spans: this
+	// Step 5a) Send a metrics message without a node but with metrics: this
 	// should be registered as belonging to the last used node i.e. "node2".
-	sLn2a := []*tracepb.Span{{TraceId: []byte("_BCDEFGHIJKLMNO_"), Status: &tracepb.Status{}}, {TraceId: []byte("_234567890abcde_"), Status: &tracepb.Status{}}}
-	err = traceClient.Send(&agenttracepb.ExportTraceServiceRequest{Node: nil, Spans: sLn2a})
+	mLn2a := []*metricspb.Metric{makeMetric(5)}
+	err = metricsClient.Send(&agentmetricspb.ExportMetricsServiceRequest{Node: nil, Metrics: mLn2a})
 	require.NoError(t, err, "Failed to send the proxied message without a node: %v", err)
 
 	// Step 5b)
-	sLn2b := []*tracepb.Span{{TraceId: []byte("_xxxxxxxxxxxxxx_"), Status: &tracepb.Status{}}, {TraceId: []byte("B234567890abcdAB"), Status: &tracepb.Status{}}}
-	err = traceClient.Send(&agenttracepb.ExportTraceServiceRequest{Node: nil, Spans: sLn2b})
+	mLn2b := []*metricspb.Metric{makeMetric(6)}
+	err = metricsClient.Send(&agentmetricspb.ExportMetricsServiceRequest{Node: nil, Metrics: mLn2b})
 	require.NoError(t, err, "Failed to send the proxied message without a node: %v", err)
 	// Give the process sometime to send data over the wire and perform batching
 	<-time.After(150 * time.Millisecond)
 
 	// Examination time!
-	resultsMapping := make(map[string][]*tracepb.Span)
-	for _, td := range spanSink.AllTraces() {
-		rss := td.ResourceSpans()
-		for i := 0; i < rss.Len(); i++ {
-			node, _, spans := internaldata.ResourceSpansToOC(rss.At(i))
-			resultsMapping[nodeToKey(node)] = append(resultsMapping[nodeToKey(node)], spans...)
+	resultsMapping := make(map[string][]*metricspb.Metric)
+	for _, md := range metricSink.AllMetrics() {
+		rms := md.ResourceMetrics()
+		for i := 0; i < rms.Len(); i++ {
+			node, _, metrics := internaldata.ResourceMetricsToOC(rms.At(i))
+			resultsMapping[nodeToKey(node)] = append(resultsMapping[nodeToKey(node)], metrics...)
 		}
 	}
 
@@ -161,71 +164,56 @@ func TestExportMultiplexing(t *testing.T) {
 		t.Errorf("Got %d keys in the results map; Wanted exactly %d\n\nResultsMapping: %+v\n", g, w, resultsMapping)
 	}
 
-	// Want span counts
-	wantSpanCounts := map[string]int{
+	// Want metric counts
+	wantMetricCounts := map[string]int{
 		nodeToKey(initiatingNode): 1,
-		nodeToKey(node1):          3,
-		nodeToKey(node2):          5,
+		nodeToKey(node1):          2,
+		nodeToKey(node2):          3,
 	}
-	for key, wantSpanCounts := range wantSpanCounts {
-		gotSpanCounts := len(resultsMapping[key])
-		if gotSpanCounts != wantSpanCounts {
-			t.Errorf("Key=%q gotSpanCounts %d wantSpanCounts %d", key, gotSpanCounts, wantSpanCounts)
+	for key, wantMetricCounts := range wantMetricCounts {
+		gotMetricCounts := len(resultsMapping[key])
+		if gotMetricCounts != wantMetricCounts {
+			t.Errorf("Key=%q gotMetricCounts %d wantMetricCounts %d", key, gotMetricCounts, wantMetricCounts)
 		}
 	}
 
-	// Now ensure that the exported spans match up exactly with
+	// Now ensure that the exported metrics match up exactly with
 	// the nodes and the last seen node expectation/behavior.
 	// (or at least their serialized equivalents match up)
-	wantContents := map[string][]*tracepb.Span{
-		nodeToKey(initiatingNode): sLi,
-		nodeToKey(node1):          append(sL1, sLn1...),
-		nodeToKey(node2):          append(sL2, append(sLn2a, sLn2b...)...),
+	wantContents := map[string][]*metricspb.Metric{
+		nodeToKey(initiatingNode): mLi,
+		nodeToKey(node1):          append(mL1, mLn1...),
+		nodeToKey(node2):          append(mL2, append(mLn2a, mLn2b...)...),
 	}
 
-	for nodeKey, wantSpans := range wantContents {
-		gotSpans, ok := resultsMapping[nodeKey]
-		if !ok {
-			t.Errorf("Wanted to find a node that was not found for key: %s", nodeKey)
-		}
-		if len(gotSpans) != len(wantSpans) {
-			t.Errorf("Unequal number of spans for nodeKey: %s", nodeKey)
-		}
-		for _, wantSpan := range wantSpans {
-			found := false
-			for _, gotSpan := range gotSpans {
-				wantStr, _ := json.Marshal(wantSpan)
-				gotStr, _ := json.Marshal(gotSpan)
-				if bytes.Equal(wantStr, gotStr) {
-					found = true
-				}
-			}
-			if !found {
-				t.Errorf("Unequal span serialization\nGot:\n\t%s\nWant:\n\t%s\n", gotSpans, wantSpans)
-			}
-		}
+	gotBlob, _ := json.Marshal(resultsMapping)
+	wantBlob, _ := json.Marshal(wantContents)
+	if !bytes.Equal(gotBlob, wantBlob) {
+		t.Errorf("Unequal serialization results\nGot:\n\t%s\nWant:\n\t%s\n", gotBlob, wantBlob)
 	}
 }
 
 // The first message without a Node MUST be rejected and teardown the connection.
 // See https://github.com/census-instrumentation/opencensus-service/issues/53
 func TestExportProtocolViolations_nodelessFirstMessage(t *testing.T) {
-	spanSink := new(consumertest.TracesSink)
+	metricSink := new(consumertest.MetricsSink)
 
-	port, doneFn := ocReceiverOnGRPCServer(t, spanSink)
+	port, doneFn := ocReceiverOnGRPCServer(t, metricSink)
 	defer doneFn()
 
-	traceClient, traceClientDoneFn, err := makeTraceServiceClient(port)
-	require.NoError(t, err, "Failed to create the gRPC TraceService_ExportClient: %v", err)
-	defer traceClientDoneFn()
+	metricsClient, metricsClientDoneFn, err := makeMetricsServiceClient(port)
+	require.NoError(t, err, "Failed to create the gRPC MetricsService_ExportClient: %v", err)
+	defer metricsClientDoneFn()
 
 	// Send a Nodeless first message
-	err = traceClient.Send(&agenttracepb.ExportTraceServiceRequest{Node: nil})
+	err = metricsClient.Send(&agentmetricspb.ExportMetricsServiceRequest{Node: nil})
 	require.NoError(t, err, "Unexpectedly failed to send the first message: %v", err)
 
 	longDuration := 2 * time.Second
-	testDone := make(chan struct{})
-	goroutineDone := make(chan struct{})
+	testDone := make(chan bool, 1)
+	var wg sync.WaitGroup
+	wg.Add(1)
+
 	go func() {
 		// Our insurance policy to ensure that this test doesn't hang
 		// forever and should quickly report if/when we regress.
@@ -233,17 +221,17 @@ func TestExportProtocolViolations_nodelessFirstMessage(t *testing.T) {
 		case <-testDone:
 			t.Log("Test ended early enough")
 		case <-time.After(longDuration):
-			traceClientDoneFn()
+			metricsClientDoneFn()
 			t.Errorf("Test took too long (%s) and is likely still hanging so this is a regression", longDuration)
 		}
-		close(goroutineDone)
+		wg.Done()
 	}()
 
 	// Now the response should return an error and should have been torn down
 	// regardless of the number of times after invocation below, or any attempt
 	// to send the proper/corrective data should be rejected.
 	for i := 0; i < 10; i++ {
-		recv, err := traceClient.Recv()
+		recv, err := metricsClient.Recv()
 		if recv != nil {
 			t.Errorf("Iteration #%d: Unexpectedly got back a response: %#v", i, recv)
 		}
@@ -263,7 +251,7 @@ func TestExportProtocolViolations_nodelessFirstMessage(t *testing.T) {
 			Identifier:  &commonpb.ProcessIdentifier{Pid: 9489, HostName: "nodejs-host"},
 			LibraryInfo: &commonpb.LibraryInfo{Language: commonpb.LibraryInfo_NODE_JS},
 		}
-		if err = traceClient.Send(&agenttracepb.ExportTraceServiceRequest{Node: n1}); err == nil {
+		if err = metricsClient.Send(&agentmetricspb.ExportMetricsServiceRequest{Node: n1}); err == nil {
 			t.Errorf("Iteration #%d: Unexpectedly succeeded in sending a message upstream. Connection must be in terminal state", i)
 		} else if g, w := err, io.EOF; g != w {
 			t.Errorf("Iteration #%d:\nGot error %q\nWant error %q", i, g, w)
@@ -271,43 +259,42 @@ func TestExportProtocolViolations_nodelessFirstMessage(t *testing.T) {
 	}
 
 	close(testDone)
-	<-goroutineDone
+	wg.Wait()
 }
 
-// If the first message is valid (has a non-nil Node) and has spans, those
-// spans should be received and NEVER discarded.
+// If the first message is valid (has a non-nil Node) and has metrics, those
+// metrics should be received and NEVER discarded.
 // See https://github.com/census-instrumentation/opencensus-service/issues/51
-func TestExportProtocolConformation_spansInFirstMessage(t *testing.T) {
-	spanSink := new(consumertest.TracesSink)
+func TestExportProtocolConformation_metricsInFirstMessage(t *testing.T) {
+	// This test used to be flaky on Windows. Skip if errors pop up again
 
-	port, doneFn := ocReceiverOnGRPCServer(t, spanSink)
+	metricSink := new(consumertest.MetricsSink)
+
+	addr, doneFn := ocReceiverOnGRPCServer(t, metricSink)
 	defer doneFn()
 
-	traceClient, traceClientDoneFn, err := makeTraceServiceClient(port)
-	require.NoError(t, err, "Failed to create the gRPC TraceService_ExportClient: %v", err)
-	defer traceClientDoneFn()
+	metricsClient, metricsClientDoneFn, err := makeMetricsServiceClient(addr)
+	require.NoError(t, err, "Failed to create the gRPC MetricsService_ExportClient: %v", err)
+	defer metricsClientDoneFn()
 
-	sLi := []*tracepb.Span{
-		{TraceId: []byte("1234567890abcdef"), Status: &tracepb.Status{}},
-		{TraceId: []byte("XXXXXXXXXXabcdef"), Status: &tracepb.Status{}},
-	}
+	mLi := []*metricspb.Metric{makeMetric(10), makeMetric(11)}
 	ni := &commonpb.Node{
 		Identifier:  &commonpb.ProcessIdentifier{Pid: 1},
 		LibraryInfo: &commonpb.LibraryInfo{Language: commonpb.LibraryInfo_JAVA},
 	}
-	err = traceClient.Send(&agenttracepb.ExportTraceServiceRequest{Node: ni, Spans: sLi})
+	err = metricsClient.Send(&agentmetricspb.ExportMetricsServiceRequest{Node: ni, Metrics: mLi})
 	require.NoError(t, err, "Failed to send the first message: %v", err)
 
 	// Give it time to be sent over the wire, then exported.
 	<-time.After(100 * time.Millisecond)
 
 	// Examination time!
-	resultsMapping := make(map[string][]*tracepb.Span)
-	for _, td := range spanSink.AllTraces() {
-		rss := td.ResourceSpans()
-		for i := 0; i < rss.Len(); i++ {
-			node, _, spans := internaldata.ResourceSpansToOC(rss.At(i))
-			resultsMapping[nodeToKey(node)] = append(resultsMapping[nodeToKey(node)], spans...)
+	resultsMapping := make(map[string][]*metricspb.Metric)
+	for _, md := range metricSink.AllMetrics() {
+		rms := md.ResourceMetrics()
+		for i := 0; i < rms.Len(); i++ {
+			node, _, metrics := internaldata.ResourceMetricsToOC(rms.At(i))
+			resultsMapping[nodeToKey(node)] = append(resultsMapping[nodeToKey(node)], metrics...)
 		}
 	}
 
@@ -322,13 +309,13 @@ func TestExportProtocolConformation_spansInFirstMessage(t *testing.T) {
 	for key, wantLength := range wantLengths {
 		gotLength := len(resultsMapping[key])
 		if gotLength != wantLength {
-			t.Errorf("Exported spans:: Key: %s\nGot length %d\nWant length %d", key, gotLength, wantLength)
+			t.Errorf("Exported metrics:: Key: %s\nGot length %d\nWant length %d", key, gotLength, wantLength)
 		}
 	}
 
 	// And finally ensure that the protos' serializations are equivalent to the expected
-	wantContents := map[string][]*tracepb.Span{
-		nodeToKey(ni): sLi,
+	wantContents := map[string][]*metricspb.Metric{
+		nodeToKey(ni): mLi,
 	}
 
 	gotBlob, _ := json.Marshal(resultsMapping)
@@ -339,21 +326,21 @@ func TestExportProtocolConformation_spansInFirstMessage(t *testing.T) {
 }
 
 // Helper functions from here on below
-func makeTraceServiceClient(addr net.Addr) (agenttracepb.TraceService_ExportClient, func(), error) {
+func makeMetricsServiceClient(addr net.Addr) (agentmetricspb.MetricsService_ExportClient, func(), error) {
 	cc, err := grpc.Dial(addr.String(), grpc.WithInsecure(), grpc.WithBlock())
 	if err != nil {
 		return nil, nil, err
 	}
 
-	svc := agenttracepb.NewTraceServiceClient(cc)
-	traceClient, err := svc.Export(context.Background())
+	svc := agentmetricspb.NewMetricsServiceClient(cc)
+	metricsClient, err := svc.Export(context.Background())
 	if err != nil {
 		_ = cc.Close()
 		return nil, nil, err
 	}
 
 	doneFn := func() { _ = cc.Close() }
-	return traceClient, doneFn, nil
+	return metricsClient, doneFn, nil
 }
 
 func nodeToKey(n *commonpb.Node) string {
@@ -361,11 +348,11 @@ func nodeToKey(n *commonpb.Node) string {
 	return string(blob)
 }
 
-func ocReceiverOnGRPCServer(t *testing.T, sr consumer.Traces) (net.Addr, func()) {
-	ln, err := net.Listen("tcp", "localhost:0")
+func ocReceiverOnGRPCServer(t *testing.T, sr consumer.Metrics) (net.Addr, func()) {
+	ln, err := net.Listen("tcp", "localhost:")
 	require.NoError(t, err, "Failed to find an available address to run the gRPC server: %v", err)
 
-	doneFnList := []func(){func() { require.NoError(t, ln.Close()) }}
+	doneFnList := []func(){func() { ln.Close() }}
 	done := func() {
 		for _, doneFn := range doneFnList {
 			doneFn()
@@ -376,13 +363,48 @@ func ocReceiverOnGRPCServer(t *testing.T, sr consumer.Traces) (net.Addr, func())
 	require.NoError(t, err, "Failed to create the Receiver: %v", err)
 
 	// Now run it as a gRPC server
-	srv := grpc.NewServer(
-		grpc.UnaryInterceptor(otelgrpc.UnaryServerInterceptor()),
-		grpc.StreamInterceptor(otelgrpc.StreamServerInterceptor()))
-	agenttracepb.RegisterTraceServiceServer(srv, oci)
+	srv := grpc.NewServer()
+	agentmetricspb.RegisterMetricsServiceServer(srv, oci)
 	go func() {
 		_ = srv.Serve(ln)
 	}()
 
 	return ln.Addr(), done
+}
+
+func makeMetric(val int) *metricspb.Metric {
+	key := &metricspb.LabelKey{
+		Key: fmt.Sprintf("%s%d", "key", val),
+	}
+	value := &metricspb.LabelValue{
+		Value:    fmt.Sprintf("%s%d", "value", val),
+		HasValue: true,
+	}
+
+	descriptor := &metricspb.MetricDescriptor{
+		Name:        fmt.Sprintf("%s%d", "metric_descriptort_", val),
+		Description: "metric descriptor",
+		Unit:        "1",
+		Type:        metricspb.MetricDescriptor_GAUGE_INT64,
+		LabelKeys:   []*metricspb.LabelKey{key},
+	}
+
+	now := time.Now().UTC()
+	point := &metricspb.Point{
+		Timestamp: timestamppb.New(now.Add(20 * time.Second)),
+		Value: &metricspb.Point_Int64Value{
+			Int64Value: int64(val),
+		},
+	}
+
+	ts := &metricspb.TimeSeries{
+		StartTimestamp: timestamppb.New(now.Add(-10 * time.Second)),
+		LabelValues:    []*metricspb.LabelValue{value},
+		Points:         []*metricspb.Point{point},
+	}
+
+	return &metricspb.Metric{
+		MetricDescriptor: descriptor,
+		Timeseries:       []*metricspb.TimeSeries{ts},
+	}
 }
